Breaking: DomEvent no longer passes event.target to the handler.
Event.on and Event.off now accept a comma or space separated list of events.
Event.on and Event.off now accept a context for the listener.
Event.on and Event.off now accept an array of listeners.
Event.on and Event.off now accept an array of tuples containing both listener and context

DomEvent.on and DomEvent.off now accept a comma or space separated list of events.
DomEvent.on and DomEvent.off now accept a context for the listener.
DomEvent.on and DomEvent.off now accept an array of listeners.
DomEvent.on and DomEvent.off now accept an array of tuples containing both listener and context